# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Federation Agricole API
  version: 2.0.0
servers:
  - url: http://localhost:8080
paths:
...
  /collectivites:
    get:
      summary: Liste des collectivités
      responses:
        200:
          description: OK
    post:
      summary: Créer une collectivité
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCollectiviteDTO'
      responses:
        201:
          description: Créée
        400:
          description: Conditions non respectées

  # ADHESION
...
  /stats/collectivite/{id}:
    get:
      summary: Statistiques collectivité
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatCollectivite'

  # STATS FEDERATION
...
components:
  schemas:
...
    CreateCollectiviteDTO:
      type: object
      properties:
        nom: { type: string }
        numero: { type: integer }
        dateCreation: { type: string, format: date }
        idVille: { type: integer }
        idDomaine: { type: integer }
        idFederation: { type: integer }
...
    StatCollectivite:
      type: object
      properties:
        nbMembres: { type: integer }
        tauxAssiduite: { type: number }
        totalCotisations: { type: number }
        impayes: { type: number }
        detailsMembres:
          type: array
          items:
            type: object
            properties:
              idMembre: { type: integer }
              tauxAssiduite: { type: number }
              totalVerse: { type: number }
              impaye: { type: number }
...